move contrib jsx to assets folder
