+++ apps/documents/assets/ChapterForm.jsx
@@ -1,6 +1,6 @@
 var React = require('react')
 var django = require('django')
-var ErrorList = require('../../contrib/static/js/ErrorList')
+var ErrorList = require('../../contrib/assets/ErrorList')
 var ParagraphForm = require('./ParagraphForm')
 var FlipMove = require('react-flip-move')
